+++ x/poolmanager/export_test.go
@@ -63,6 +63,14 @@ func (k Keeper) CreateOsmoMultihopExpectedSwapOuts(
 	return k.createOsmoMultihopExpectedSwapOuts(ctx, route, tokenOut, cumulativeRouteSwapFee, sumOfSwapFees)
 }
 
+func (k Keeper) CalcTakerFeeExactIn(tokenIn sdk.Coin, takerFee sdk.Dec) (sdk.Coin, sdk.Coin) {
+	return k.calcTakerFeeExactIn(tokenIn, takerFee)
+}
+
+func (k Keeper) CalcTakerFeeExactOut(tokenOut sdk.Coin, takerFee sdk.Dec) (sdk.Coin, sdk.Coin) {
+	return k.calcTakerFeeExactOut(tokenOut, takerFee)
+}
+
 func (k Keeper) TrackVolume(ctx sdk.Context, poolId uint64, volumeGenerated sdk.Coin) {
 	k.trackVolume(ctx, poolId, volumeGenerated)
 }