+++ x/txfees/keeper/hooks.go
@@ -14,36 +14,32 @@ func (k Keeper) BeforeEpochStart(ctx sdk.Context, epochIdentifier string, epochN
 
 // at the end of each epoch, swap all non-OSMO fees into OSMO and transfer to fee module account
 func (k Keeper) AfterEpochEnd(ctx sdk.Context, epochIdentifier string, epochNumber int64) error {
-	nonNativeFeeAddr := k.accountKeeper.GetModuleAddress(txfeestypes.NonNativeFeeCollectorForStakingRewardsName)
-	baseDenom, _ := k.GetBaseDenom(ctx)
-	feeTokens := k.GetFeeTokens(ctx)
+	defaultFeesDenom, _ := k.GetBaseDenom(ctx)
 
-	for _, feetoken := range feeTokens {
-		if feetoken.Denom == baseDenom {
-			continue
-		}
-		coinBalance := k.bankKeeper.GetBalance(ctx, nonNativeFeeAddr, feetoken.Denom)
-		if coinBalance.Amount.IsZero() {
-			continue
-		}
+	nonNativeStakingCollectorAddress := k.accountKeeper.GetModuleAddress(txfeestypes.FeeCollectorForStakingRewardsName)
 
-		// Do the swap of this fee token denom to base denom.
-		_ = osmoutils.ApplyFuncIfNoError(ctx, func(cacheCtx sdk.Context) error {
-			// We allow full slippage. Theres not really an effective way to bound slippage until TWAP's land,
-			// but even then the point is a bit moot.
-			// The only thing that could be done is a costly griefing attack to reduce the amount of osmo given as tx fees.
-			// However the idea of the txfees FeeToken gating is that the pool is sufficiently liquid for that base token.
-			minAmountOut := sdk.ZeroInt()
-			_, err := k.poolManager.SwapExactAmountIn(cacheCtx, nonNativeFeeAddr, feetoken.PoolID, coinBalance, baseDenom, minAmountOut)
-			return err
-		})
-	}
+	// Non-native fee collector for staking rewards get swapped entirely into base denom.
+	k.swapNonNativeFeeToDenom(ctx, defaultFeesDenom, nonNativeStakingCollectorAddress)
 
-	// Get all of the txfee payout denom in the module account
-	baseDenomCoins := sdk.NewCoins(k.bankKeeper.GetBalance(ctx, nonNativeFeeAddr, baseDenom))
-
+	// Now that the rewards have been swapped, transfer any base denom existing in the non-native fee collector to the fee collector (indirectly distributing to stakers)
+	baseDenomCoins := sdk.NewCoins(k.bankKeeper.GetBalance(ctx, nonNativeStakingCollectorAddress, defaultFeesDenom))
 	_ = osmoutils.ApplyFuncIfNoError(ctx, func(cacheCtx sdk.Context) error {
-		err := k.bankKeeper.SendCoinsFromModuleToModule(ctx, txfeestypes.NonNativeFeeCollectorForStakingRewardsName, txfeestypes.FeeCollectorName, baseDenomCoins)
+		err := k.bankKeeper.SendCoinsFromModuleToModule(ctx, txfeestypes.FeeCollectorForStakingRewardsName, txfeestypes.FeeCollectorName, baseDenomCoins)
+		return err
+	})
+
+	// Non-native fee collector for community pool get swapped entirely into denom specified in the pool manager params.
+	poolManagerParams := k.poolManager.GetParams(ctx)
+	denomToSwapTo := poolManagerParams.TakerFeeParams.CommunityPoolDenomToSwapNonWhitelistedAssetsTo
+	// Only non-whitelisted assets should exist here since we do direct community pool funds when calculating the taker fee if
+	// the input is a whitelisted asset.
+	nonNativeCommunityPoolCollectorAddress := k.accountKeeper.GetModuleAddress(txfeestypes.FeeCollectorForCommunityPoolName)
+	k.swapNonNativeFeeToDenom(ctx, denomToSwapTo, nonNativeCommunityPoolCollectorAddress)
+
+	// Now that the non whitelisted assets have been swapped, fund the community pool with the denom we swapped to.
+	denomToSwapToCoins := sdk.NewCoins(k.bankKeeper.GetBalance(ctx, nonNativeCommunityPoolCollectorAddress, denomToSwapTo))
+	_ = osmoutils.ApplyFuncIfNoError(ctx, func(cacheCtx sdk.Context) error {
+		err := k.distributionKeeper.FundCommunityPool(ctx, denomToSwapToCoins, nonNativeCommunityPoolCollectorAddress)
 		return err
 	})
 
@@ -68,4 +64,48 @@ func (h Hooks) BeforeEpochStart(ctx sdk.Context, epochIdentifier string, epochNu
 
 func (h Hooks) AfterEpochEnd(ctx sdk.Context, epochIdentifier string, epochNumber int64) error {
 	return h.k.AfterEpochEnd(ctx, epochIdentifier, epochNumber)
+}
+
+// swapNonNativeFeeToDenom swaps the given non-native fees into the given denom from the given fee collector address.
+// If an error in swap occurs for a given denom, it will be silently skipped.
+// CONTRACT: a pool must exist between each denom in the balance and denomToSwapTo. If doesn't exist. Silently skip swap.
+// CONTRACT: protorev must be configured to have a pool for the given denom pair. Otherwise, the denom will be skipped.
+func (k Keeper) swapNonNativeFeeToDenom(ctx sdk.Context, denomToSwapTo string, feeCollectorAddress sdk.AccAddress) {
+	feeCollectorBalance := k.bankKeeper.GetAllBalances(ctx, feeCollectorAddress)
+
+	for _, coin := range feeCollectorBalance {
+		if coin.Denom == denomToSwapTo {
+			continue
+		}
+
+		// Search for the denom pair route via the protorev store.
+		// Since OSMO is one of the protorev denoms, many of the routes will exist in this store.
+		// There will be times when this store does not know about a route, but this is acceptable
+		// since this will likely be a very small value of a relatively unknown token. If this begins
+		// to accrue more value, we can always manually register the route and it will get swapped in
+		// the next epoch.
+		poolId, err := k.protorevKeeper.GetPoolForDenomPairNoOrder(ctx, denomToSwapTo, coin.Denom)
+		if err != nil {
+			if err != nil {
+				// The pool route either doesn't exist or is disabled in protorev.
+				// It will just accrue in the non-native fee collector account.
+				// Skip this denom and move on to the next one.
+				continue
+			}
+		}
+
+		// Do the swap of this fee token denom to base denom.
+		_ = osmoutils.ApplyFuncIfNoError(ctx, func(cacheCtx sdk.Context) error {
+			// We allow full slippage. Theres not really an effective way to bound slippage until TWAP's land,
+			// but even then the point is a bit moot.
+			// The only thing that could be done is a costly griefing attack to reduce the amount of osmo given as tx fees.
+			// However the idea of the txfees FeeToken gating is that the pool is sufficiently liquid for that base token.
+			minAmountOut := sdk.ZeroInt()
+
+			// We swap without charging a taker fee / sending to the non native fee collector, since these are funds that
+			// are accruing from the taker fee itself.
+			_, err := k.poolManager.SwapExactAmountInNoTakerFee(cacheCtx, feeCollectorAddress, poolId, coin, denomToSwapTo, minAmountOut)
+			return err
+		})
+	}
 }