+++ x/txfees/keeper/keeper_test.go
@@ -11,6 +11,7 @@ import (
 	osmosisapp "github.com/osmosis-labs/osmosis/v19/app"
 
 	"github.com/osmosis-labs/osmosis/v19/app/apptesting"
+	protorevtypes "github.com/osmosis-labs/osmosis/v19/x/protorev/types"
 	"github.com/osmosis-labs/osmosis/v19/x/txfees/types"
 )
 
@@ -35,6 +36,21 @@ func (s *KeeperTestSuite) SetupTest(isCheckTx bool) {
 		WithTxConfig(encodingConfig.TxConfig).
 		WithLegacyAmino(encodingConfig.Amino).
 		WithCodec(encodingConfig.Marshaler)
+
+	// We set the base denom here in order for highest liquidity routes to get generated.
+	// This is used in the tx fees epoch hook to swap the non OSMO to other tokens.
+	baseDenom, err := s.App.TxFeesKeeper.GetBaseDenom(s.Ctx)
+	s.Require().NoError(err)
+
+	// Configure protorev base denoms
+	baseDenomPriorities := []protorevtypes.BaseDenom{
+		{
+			Denom:    baseDenom,
+			StepSize: sdk.NewInt(1_000_000),
+		},
+	}
+	err = s.App.ProtoRevKeeper.SetBaseDenoms(s.Ctx, baseDenomPriorities)
+	s.Require().NoError(err)
 
 	// Mint some assets to the accounts.
 	for _, acc := range s.TestAccs {