+++ x/txfees/types/expected_keepers.go
@@ -31,14 +31,14 @@ type PoolManager interface {
 		tokenOutMinAmount sdk.Int,
 	) (sdk.Int, error)
 
-	// SwapExactAmountInNoTakerFee(
-	// 	ctx sdk.Context,
-	// 	sender sdk.AccAddress,
-	// 	poolId uint64,
-	// 	tokenIn sdk.Coin,
-	// 	tokenOutDenom string,
-	// 	tokenOutMinAmount sdk.Int,
-	// ) (sdk.Int, error)
+	SwapExactAmountInNoTakerFee(
+		ctx sdk.Context,
+		sender sdk.AccAddress,
+		poolId uint64,
+		tokenIn sdk.Coin,
+		tokenOutDenom string,
+		tokenOutMinAmount sdk.Int,
+	) (sdk.Int, error)
 
 	GetParams(ctx sdk.Context) (params poolmanagertypes.Params)
 }